team page breaks when zero members; addresses #216
The team page breaks when there are zero members. Add “if members” checks in the jade file. This solves the issue. In the future we might consider updating the methods in the models to handle cases where there are zero members. Fixing the jade file was the simplest and safest fix for now.